+++ ChangeLog.md
@@ -77,6 +77,7 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 -  `@moxb/meteor` is now using the global `Meteor` and `Tracker` variables, because `meteor/meteor` is not a real
    package and therefore there are some problems with dependencies....
 
+- rename `extractErrorMessage` to `extractErrorMessages` and added `extractErrorString` function
 ### Removed
 
 ### Fixed